+++ src/settings/Locale.js
@@ -1,65 +1,23 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { FormattedMessage, injectIntl, createIntl, createIntlCache } from 'react-intl';
-import { Field } from 'redux-form';
-
-import { IfPermission, supportedLocales } from '@folio/stripes/core';
 import { ConfigManager } from '@folio/stripes/smart-components';
-import { Button, Col, Row, Select, CurrencySelect } from '@folio/stripes/components';
-import timezones from '../util/timezones';
 
-const timeZonesList = timezones.map(timezone => (
-  {
-    value: timezone.value,
-    label: timezone.value,
-  }
-));
-
-const options = supportedLocales.map(k => ({ value: k, label: '' }));
+import LocaleForm from './LocaleForm';
 
 class Locale extends React.Component {
   static propTypes = {
     stripes: PropTypes.shape({
-      connect: PropTypes.func.isRequired,
-      logger: PropTypes.shape({
-        log: PropTypes.func.isRequired,
-      }).isRequired,
+      setCurrency: PropTypes.func.isRequired,
       setLocale: PropTypes.func.isRequired,
       setTimezone: PropTypes.func.isRequired,
-      setCurrency: PropTypes.func.isRequired,
+      connect: PropTypes.func.isRequired,
     }).isRequired,
     label: PropTypes.node.isRequired,
-    intl: PropTypes.object.isRequired,
   };
 
   constructor(props) {
     super(props);
-
     this.configManager = props.stripes.connect(ConfigManager);
-    this.setLocaleSettings = this.setLocaleSettings.bind(this);
-    this.getInitialValues = this.getInitialValues.bind(this);
-    this.beforeSave = this.beforeSave.bind(this);
-
-    // This is optional but highly recommended
-    // since it prevents memory leak
-    const cache = createIntlCache();
-    const { intl } = props;
-
-    options.forEach(locale => {
-      locale.intl = createIntl({
-        locale: locale.value,
-        messages: {}
-      }, cache);
-
-      // label contains language in current locale and in destination locale
-      // e.g. given the current locale is `en` and the keys `ar` and `zh-CN` show:
-      //     Arabic / العربية
-      //     Chinese (China) / 中文（中国）
-      // e.g. given the current locale is `ar` and the keys `ar` and `zh-CN` show:
-      //    العربية / العربية
-      //    الصينية (الصين) / 中文（中国）
-      locale.label = `${intl.formatDisplayName(locale.value, { type: 'language' })} / ${locale.intl.formatDisplayName(locale.value, { type: 'language' })}`;
-    });
   }
 
   // eslint-disable-next-line class-methods-use-this
@@ -70,6 +28,19 @@ class Locale extends React.Component {
 
     try {
       config = Object.assign({}, defaultConfig, JSON.parse(value));
+
+      // a numbering system may be glommed onto the locale.
+      // separate it if so, allowing the locale and numbering system
+      // to be configured independently. this is pretty brain-dead
+      // right now and proper locale-value parsing will need to be
+      // implemented if we choose to support for additional options.
+      if (config.locale) {
+        const parts = config.locale.split('-u-nu-');
+        config.locale = parts[0];
+        if (parts[1]) {
+          config.numberingSystem = parts[0];
+        }
+      }
     } catch (e) {
       config = defaultConfig;
     }
@@ -77,20 +48,29 @@ class Locale extends React.Component {
     return config;
   }
 
-  setLocaleSettings(setting) {
+  afterSave(setting) {
     const localeValues = JSON.parse(setting.value);
-    const { locale, timezone, currency } = localeValues;
+    let { locale } = localeValues;
+    const { numberingSystem, timezone, currency } = localeValues;
+
+    // A numbering system other than the locale's default
+    // may be configured with `-u-nu-SYSTEM`
+    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Intl/Locale/numberingSystem
+    if (numberingSystem) {
+      locale = `${locale}-u-nu-${numberingSystem}`;
+    }
+
     setTimeout(() => {
       if (locale) this.props.stripes.setLocale(locale);
       if (timezone) this.props.stripes.setTimezone(timezone);
       if (currency) this.props.stripes.setCurrency(currency);
-    }, 2000);
+    }, 500);
   }
 
   // eslint-disable-next-line class-methods-use-this
   beforeSave(data) {
-    const { locale, timezone, currency } = data;
-    return JSON.stringify({ locale, timezone, currency });
+    const { locale, numberingSystem, timezone, currency } = data;
+    return JSON.stringify({ locale, numberingSystem, timezone, currency });
   }
 
   render() {
@@ -98,75 +78,14 @@ class Locale extends React.Component {
       <this.configManager
         label={this.props.label}
         moduleName="ORG"
-        getInitialValues={this.getInitialValues}
         configName="localeSettings"
         onBeforeSave={this.beforeSave}
-        onAfterSave={this.setLocaleSettings}
-      >
-        <IfPermission perm="ui-developer.settings.locale">
-          <Row>
-            <Col xs={12}>
-              <p>
-                <FormattedMessage id="ui-tenant-settings.settings.locale.localeWarning" values={{ label: <FormattedMessage id="ui-tenant-settings.settings.locale.changeSessionLocale" /> }} />
-              </p>
-              <div>
-                <Button to="/settings/developer/locale">
-                  <FormattedMessage id="ui-tenant-settings.settings.locale.changeSessionLocale" />
-                </Button>
-              </div>
-            </Col>
-          </Row>
-        </IfPermission>
-        <Row>
-          <Col xs={12} id="select-locale">
-            <FormattedMessage id="ui-tenant-settings.settings.localization">
-              {label => (
-                <Field
-                  component={Select}
-                  id="locale"
-                  name="locale"
-                  placeholder="---"
-                  dataOptions={options}
-                  label={label}
-                />
-              )}
-            </FormattedMessage>
-          </Col>
-        </Row>
-        <Row>
-          <Col xs={12} id="select-timezone">
-            <FormattedMessage id="ui-tenant-settings.settings.timeZonePicker">
-              {label => (
-                <Field
-                  component={Select}
-                  id="timezone"
-                  name="timezone"
-                  placeholder="---"
-                  dataOptions={timeZonesList}
-                  label={label}
-                />
-              )}
-            </FormattedMessage>
-          </Col>
-        </Row>
-        <Row>
-          <Col xs={12} id="select-currency">
-            <FormattedMessage id="ui-tenant-settings.settings.primaryCurrency">
-              {label => (
-                <Field
-                  component={CurrencySelect}
-                  id="currency"
-                  name="currency"
-                  placeholder="---"
-                  label={label}
-                />
-              )}
-            </FormattedMessage>
-          </Col>
-        </Row>
-      </this.configManager>
+        onAfterSave={this.afterSave}
+        configFormComponent={LocaleForm}
+        getInitialValues={this.getInitialValues}
+      />
     );
   }
 }
 
-export default injectIntl(Locale);
+export default Locale;