+++ test/jest/__mocks__/currencyData.mock.js
@@ -1,1 +1,10 @@
-jest.mock('currency-codes/data', () => ({ filter: () => [] }));
+jest.mock('currency-codes/data', () => ({ filter: () => [
+  {
+    code: 'EUR',
+    currency: 'Euro',
+  },
+  {
+    code: 'USD',
+    currency: 'US Dollar',
+  },
+] }));