+++ test/jest/__mocks__/stripesSmartComponents.mock.js
@@ -4,4 +4,8 @@ jest.mock('@folio/stripes/smart-components', () => ({
   ...jest.requireActual('@folio/stripes/smart-components'),
   LocationLookup: () => <div>LocationLookup</div>,
   ViewMetaData: () => <div>ViewMetaData</div>,
-}), { virtual: true });
+  // ConfigManager: jest.fn(({ children }) => <div>{children}</div>),
+  // ConfigManager: ({ children }) => <div>{children}</div>,
+  // ConfigManager: () => <div>Config manager</div>,
+// }), { virtual: true });
+}));